@@ -55,25 +55,35 @@
     "email": "peter.han.9493@gmail.com",
     "url": "https://github.com/peterthehan"
   },
-  "main": "./src/index.js",
+  "main": "./dist/src/index.js",
   "bin": {
-    "runescape-text": "./src/cli.js"
+    "runescape-text": "./dist/src/cli.js"
   },
   "files": [
-    "src/**"
+    "dist/**"
   ],
   "scripts": {
-    "format": "npx eslint --fix . && prettier --no-config --write .",
-    "start": "node src/cli.js",
-    "test": "jest"
+    "build": "rm -rf ./dist && npx tsc && cp -r ./src/assets ./dist/src",
+    "lint": "npx eslint --fix .",
+    "start": "node src/cli.js"
   },
   "dependencies": {
-    "canvas": "^2.11.2",
+    "canvas": "^2.8.0",
     "gif-encoder": "^0.7.2",
     "word-wrap": "^1.2.3"
   },
   "devDependencies": {
-    "eslint": "^7.27.0",
-    "prettier": "^2.3.0"
+    "@tsconfig/recommended": "^1.0.2",
+    "@types/gif-encoder": "^0.7.1",
+    "@types/jest": "^29.5.2",
+    "@typescript-eslint/eslint-plugin": "^5.59.8",
+    "@typescript-eslint/parser": "^5.59.8",
+    "eslint": "^8.41.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "jest": "^29.5.0",
+    "prettier": "^2.8.8",
+    "ts-jest": "^29.1.0",
+    "typescript": "^5.0.4"
   }
 }
