@@ -1,5 +1,5 @@
 class InvalidArgumentError extends Error {
-  constructor(message) {
+  constructor(message: string) {
     super(message);
     this.name = "InvalidArgumentError";
   }
