@@ -37,11 +37,21 @@ const flash1Osrs = [red, yellow];
 const flash2Osrs = [cyan, blue];
 const flash3Osrs = [darkGreen, lightGreen];
 
-const isAnimated = (color, motion) => {
-  return motion !== "none" || dynamicColors.includes(color);
+const isAnimated = (color: Color, motion: Motion) => {
+  return (
+    motion !== Motion.None ||
+    [
+      Color.Glow1,
+      Color.Glow2,
+      Color.Glow3,
+      Color.Flash1,
+      Color.Flash2,
+      Color.Flash3,
+    ].includes(color)
+  );
 };
 
-module.exports = {
+export {
   colors,
   motions,
   effectsMap,
